requirements.txt
Authlib==1.5.1
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
gspread==6.2.1
oauthlib==3.3.1
pandas==2.3.2
pypdf==6.1.0
reportlab==4.4.4
streamlit==1.50.0
tzdata==2025.2
